feat(input): real keyboard/pointer control state — Xlib13 97→115 PASS (+18)
Replace the ChangeKeyboardControl/Bell/ChangePointerControl log_void
stubs with real state, mirroring Xorg dix/devices.c semantics:

- KeyboardControlState on ServerState: key click, bell percent/pitch/
  duration, LED mask, per-key auto-repeat bitmap (Xorg
  DEFAULT_AUTOREPEATS), global auto-repeat. Validation per
  DoChangeKeyboardControl: -1 restores defaults, BadValue on range,
  BadMatch for led-without-led-mode and key-without-auto-repeat-mode,
  staged commit (no partial state on error).
- PointerControlState: accel numerator/denominator + threshold with
  ProcChangePointerControl validation (denominator <= 0 → BadValue).
- GetKeyboardControl / GetPointerControl now reflect stored state
  instead of hardcoded values.
- Bell validates percent (-100..=100) → BadValue.
- Key auto-repeat machinery now honors the control state: global off
  or per-key bit cleared disarms/suppresses repeat (run.rs gate at
  arm + fire).

Also fixes two pre-existing clippy warnings (useless u32::from in
request_lengths.rs, collapsible if in process_disconnect.rs).

XTS Xlib13 (vng): XChangeKeyboardControl 6→16 PASS,
XChangePointerControl 6→11, XGetKeyboardControl/XGetPointerControl
0→1 each, XAutoRepeatOff 0→1. Note: 48 yserver-core lib test
failures are pre-existing at HEAD (verified via stash), XI property
domain.

+++ crates/yserver-protocol/src/x11/request_lengths.rs
@@ -305,7 +305,7 @@ pub fn xi_exact_required_length(minor: u8, body: &[u8]) -> Option<u32> {
         //   count(u16) num_events(u8) pad×3 events(32 × num_events)
         //   classes(u32 × count).
         31 if body.len() >= 9 => {
-            let class_count = u32::from(read_u16_le(&body[6..8]));
+            let class_count = read_u16_le(&body[6..8]);
             let num_events = u32::from(body[8]);
             Some(4 + 8 * num_events + class_count)
         }
